[nrf fromlist] Don't rely on the target config system in the nrf WDT driver
We are having some problems getting the target configuration system to
pass on config values into the driver.

We would prefer to simply encode the information we need straight into
the C driver, instead of relying on a complex config system that we
are unfamiliar with.

This makes porting to new systems simpler as developers only need to
make changes to C code to support new targets.

This follows prior art and aligns the UART and WDT drivers, as we
don't use the target config system to find the UART peripheral base
address either.

Note that on nordic platforms we don't support a timeout value other
than 500000000 us.

[nrf fromlist] because this was cherry-picked from
https://github.com/ARM-software/psa-arch-tests/pull/367
cc20d2e1adc43dd954af27c94f30531d0e91cb8f

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>